- removed google drive support, + add sftp support but without functionality, + add statusbar to backup form, + removed bugg that automatic action code hasnt the model id as the parameter (was always None)

+++ automatic_backup/models/models.py
@@ -9,9 +9,6 @@ import botocore
 import dropbox
 from dropbox import exceptions as dropbox_exceptions
 
-from pydrive.auth import GoogleAuth
-from pydrive.drive import GoogleDrive
-
 import owncloud
 
 import os
@@ -22,44 +19,42 @@ import datetime
 class BackupTypes(Enum):
     s3 = ('s3', 'Amazon Web-Service S3')
     dropbox = ('dropbox', 'Dropbox')
-    google_drive = ('google_drive', 'Google Drive')
     owncloud = ('owncloud', 'OwnCloud')
+    sftp = ('sftp', 'SFTP')
 
 
 class Configuration(models.Model):
     _name = 'automatic_backup.configuration'
     _inherit = ['mail.thread']
 
-    name = fields.Char(required=True)
-    active = fields.Boolean(default=True)
+    name = fields.Char(required=1)
+    active = fields.Boolean(default=0)
+    state = fields.Selection([('disabled', 'Disabled'),('active', 'Active')], default='disabled', readonly=1)
 
     schedule_frequently = fields.Selection(string='Every', selection=[('days', 'Days'), ('weeks', 'Weeks')],
                                            default='weeks', required=1)
     schedule_number = fields.Integer(min=1, default=1, required=1)
 
-    next_backup_time = fields.Datetime(store=0, required=1, readonly=1, compute='_compute_next_backup_time')
+    next_backup_time = fields.Datetime(store=0, readonly=1, compute='_compute_next_backup_time')
     success_mail = fields.Many2one('res.users', ondelete='set null')
     error_mail = fields.Many2one('res.users', ondelete='set null')
 
     cron_id = fields.Many2one('ir.cron', ondelete='cascade')
 
     backup_type = fields.Selection([
-        BackupTypes.s3.value,
-        BackupTypes.dropbox.value,
-        BackupTypes.google_drive.value,
-        BackupTypes.owncloud.value
-    ], required=True)
+        BackupTypes.s3.value, BackupTypes.dropbox.value, BackupTypes.owncloud.value, BackupTypes.sftp.value
+    ], required=1)
     upload_path = fields.Char("Path to upload", required=True)
-    last_backup = fields.Datetime(readonly=True)
-    last_message = fields.Char(readonly=True)
-    last_path = fields.Char(readonly=True)
+    last_backup = fields.Datetime(readonly=1)
+    last_message = fields.Char(readonly=1)
+    last_path = fields.Char(readonly=1)
 
-    show_s3 = fields.Boolean(compute='set_show_s3', store=False)
-    show_dropbox = fields.Boolean(compute='set_show_dropbox', store=False)
-    show_owncloud = fields.Boolean(compute='set_show_owncloud', store=False)
+    show_s3 = fields.Boolean(compute='set_show_s3', store=0)
+    show_dropbox = fields.Boolean(compute='set_show_dropbox', store=0)
+    show_owncloud = fields.Boolean(compute='set_show_owncloud', store=0)
 
-    show_access_key = fields.Boolean(compute='set_show_access_key', store=False)
-    show_secret_key = fields.Boolean(compute='set_show_secret_key', store=False)
+    show_access_key = fields.Boolean(compute='set_show_access_key', store=0)
+    show_secret_key = fields.Boolean(compute='set_show_secret_key', store=0)
 
     access_key_id = fields.Char("Access Key")
     secret_access_key = fields.Char("Secret Key")
@@ -72,24 +67,6 @@ class Configuration(models.Model):
     @api.model
     def create(self, vals):
         vals = self.check_upload_path(vals)
-        model_id = self.env['ir.model'].search([('model','=', self._name)])
-        nextbackuptime = self.next_backup_time
-        if not nextbackuptime:
-            nextbackuptime = datetime.datetime.today() + datetime.timedelta(days=1)
-        cron_id = self.env['ir.cron'].create({
-            'name': 'Backup: ' + vals['name'],
-            'model_id': model_id.id,
-            'state': 'code',
-            'user_id': 1,
-            'interval_number': vals['schedule_number'],
-            'interval_type': vals['schedule_frequently'],
-            'nextcall': nextbackuptime,
-            'priority': 100,
-            'numbercall': -1,
-            'active': vals['active'],
-            'code': 'model.action_backup('+str(self.id)+')'
-        })
-        vals['cron_id'] = cron_id.id
         result = super(Configuration, self).create(vals)
         return result
 
@@ -109,14 +86,18 @@ class Configuration(models.Model):
                 if 'name' in vals:
                     self.cron_id.write({'name': 'Backup: '+self.name})
         except:
+            # only if cron is busy
             pass
         return result
 
+    @api.multi
     def unlink(self):
-        cron_id = self.cron_id
-        self.cron_id = False
-        if cron_id:
-            cron_id.unlink()
+        self.ensure_one()
+        if self.cron_id:
+            cron_id = self.cron_id
+            self.cron_id = False
+            if cron_id:
+                cron_id.unlink()
         result = super(Configuration, self).unlink()
         return result
 
@@ -166,7 +147,7 @@ class Configuration(models.Model):
         self.set_show_owncloud()
 
     def action_backup(self, id):
-        backup_ids = self.browse([id])
+        backup_ids = self.search([id])
         for backup in backup_ids:
             backup.btn_action_backup()
 
@@ -178,10 +159,10 @@ class Configuration(models.Model):
                 self._backup_on_s3()
             elif self.backup_type == BackupTypes.dropbox.value[0]:
                 self._backup_on_dropbox()
-            elif self.backup_type == BackupTypes.google_drive.value[0]:
-                self._backup_on_googledrive()
             elif self.backup_type == BackupTypes.owncloud.value[0]:
                 self._backup_on_owncloud()
+            elif self.backup_type == BackupTypes.sftp.value[0]:
+                self._backup_on_sftp()
             self.send_email()
         except Exception as err:
             self.set_last_fields(err.args[0])
@@ -226,9 +207,6 @@ class Configuration(models.Model):
             self.set_last_fields(message, path=path)
             self.message_post(message)
 
-    def _backup_on_googledrive(self):
-        pass
-
     def _backup_on_owncloud(self):
         oc = owncloud.Client(self.cloud_url)
         try:
@@ -245,6 +223,9 @@ class Configuration(models.Model):
         message = ("ownCloud: No Error during upload. You can find the backup under {0}".format(path))
         self.set_last_fields(message, path=path)
         self.message_post(message)
+
+    def _backup_on_sftp(self):
+        pass
 
     def get_path_and_content(self):
         filename, content = self.get_backup()
@@ -264,3 +245,31 @@ class Configuration(models.Model):
             self.last_path = path
             self.last_backup = datetime.datetime.now()
         self.last_message = message
+
+    def deactivate_progressbar(self):
+        self.active = False
+        self.state = 'disabled'
+
+    def activate_progressbar(self):
+        self.active = True
+        self.state = 'active'
+
+        if not self.cron_id:
+            nextbackuptime = self.next_backup_time
+            if not nextbackuptime:
+                nextbackuptime = datetime.datetime.today() + datetime.timedelta(days=1)
+            model_id = self.env['ir.model'].search([('model', '=', self._name)])
+            cron_id = self.env['ir.cron'].create({
+                'name': 'Backup: ' + self.name,
+                'model_id': model_id.id,
+                'state': 'code',
+                'user_id': 1,
+                'interval_number': self.schedule_number,
+                'interval_type': self.schedule_frequently,
+                'nextcall': nextbackuptime,
+                'priority': 100,
+                'numbercall': -1,
+                'active': self.active,
+                'code': 'model.action_backup('+str(self.id)+')'
+            })
+            self.cron_id = cron_id.id
